requirements.txt
scipy==1.5.2
cleverhans==3.0.1
PyWavelets==1.1.1
matplotlib==3.3.1
tensorflow==1.14.0
scikit_image==0.16.2
Keras==2.2.4
joblib==0.16.0
numpy==1.19.1
ISR==2.2.0
Pillow==8.3.0
resnet==0.1
skimage==0.0
